Add unit pathfinding, targeting, and combat
